// src/app/shared/data-access/models/studentDocument.models.ts
export interface StudentDocument {
  adminDocumentId: number;
  amount: number;
  applicationId: number;
  bbdDescription: string;
  documentBlobName: string;
  documentTypeId: number;
  expenseCategoryId: number;
  isDeleted: boolean;
  status: number;
  statusId: number;
  universityDescription: number;
}